requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
Authlib==1.3.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-health-checking==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.6
langchain-community==0.3.4
langchain-core==0.3.15
langchain-huggingface==0.1.2
langchain-openai==0.2.5
langchain-text-splitters==0.3.1
langsmith==0.1.139
llama-cloud==0.1.4
llama-index==0.11.20
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.20
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-ollama==0.3.4
llama-index-llms-openai==0.2.16
llama-index-llms-replicate==0.3.0
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.12
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
ollama==0.3.3
openai==1.53.0
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
protobuf==5.28.3
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tk==0.1.0
tokenizers==0.20.1
torch==2.5.1
tqdm==4.66.6
transformers==4.46.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
validators==0.34.0
watchfiles==0.24.0
weaviate-client==4.9.2
websockets==13.1
wrapt==1.16.0
yarl==1.17.1
